ADMIN ALMOST DONE

+++ admin/delete-category.php
@@ -6,12 +6,16 @@ if (isset($_GET['id'])) {
     $id = filter_var($_GET['id'], FILTER_SANITIZE_NUMBER_INT);
 
 
-    $update_query = "UPDATE posts SET category_id=0 WHERE category_id=$id";
+    $update_query = "UPDATE posts SET category_id=9 WHERE category_id=$id";
     $update_result = mysqli_query($connection, $update_query);
 
-    $query = "DELETE FROM categories WHERE id=$id LIMIT 1";
-    $result = mysqli_query($connection, $query);
-    $_SESSION['delete-category-success'] = "Successfully deleted that category";
+    if (!mysqli_errno($connection)) {
+
+        $query = "DELETE FROM categories WHERE id=$id LIMIT 1";
+        $result = mysqli_query($connection, $query);
+        $_SESSION['delete-category-success'] = "Successfully deleted that category";
+
+    }
 }
 
 header('location: ' . ROOT_URL . 'admin/manage-categories.php');
